pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "ipython>=9.11.0",
    "lancedb>=0.29.2",
    "langchain",
    "langchain-community>=0.4.1",
    "langchain-openai",
    "langgraph",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.2",
    "python-socks>=2.8.1",
    "redis>=7.3.0",
    "shortuuid>=1.0.13",
    "tantivy>=0.25.1",
    "websockets>=16.0",
]
